Remove all remaining explicit <br>s
At this point, I'm confortable letting all line breaks be "soft",
to the extent that they're all part of the encoding.
How closely these encoded line breaks will be followed is ultimately
up to the end consumer application: if they want to fold whitespace,
they might as well.

A future update to the schema may add a field with a hint for cards
which are particularly sensitive to line breaking, but ultimately,
all cards have a certain amount of "comic timing" inherent to their
selected break points, regardless of font concerns. (If you look at
some of the layouts / flows of text on official Cards Against Humanity
cards, they don't always flow words onto the rest of the line, even
when there's clearly space available, when the line break would
interrupt the flow of reading the text on the card).